moved the add to cart logic to separated file
