+++ index.html
@@ -321,10 +321,10 @@
     </section>
     <!-- footer section ends -->
 
-    <!-- custom js file link  -->
-    <script src="js/script.js"></script>
-
     <!-- supplements items provider -->
     <script type="module" src="./js/SupplementsProvider.js"></script>
+
+    <!-- custom js file link  -->
+    <script src="js/script.js"></script>
   </body>
 </html>
